- Clean up of static usage of ApplicationInfoManager in DiscoveryClient to avoid race condition in references. - Added tests for eventBus events
