+++ eureka-client/src/test/java/com/netflix/discovery/DiscoveryClientRegisterUpdateTest.java
@@ -37,7 +37,7 @@ public class DiscoveryClientRegisterUpdateTest {
 
         applicationInfoManager = new TestApplicationInfoManager();
         applicationInfoManager.initComponent(new MyDataCenterInstanceConfig());
-        client = new DiscoveryClient(applicationInfoManager.getInfo(), new DefaultEurekaClientConfig());
+        client = new DiscoveryClient(applicationInfoManager, new DefaultEurekaClientConfig());
     }
 
     @After
