+++ eureka-examples/src/main/java/com/netflix/eureka/ExampleEurekaGovernatedService.java
@@ -2,12 +2,16 @@ package com.netflix.eureka;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Injector;
+import com.google.inject.Scopes;
 import com.netflix.appinfo.EurekaInstanceConfig;
 import com.netflix.appinfo.MyDataCenterInstanceConfig;
 import com.netflix.config.DynamicPropertyFactory;
 import com.netflix.discovery.guice.EurekaModule;
+import com.netflix.eventbus.impl.EventBusImpl;
+import com.netflix.eventbus.spi.EventBus;
 import com.netflix.governator.guice.LifecycleInjector;
 import com.netflix.governator.guice.LifecycleInjectorBuilder;
+import com.netflix.governator.guice.LifecycleInjectorMode;
 import com.netflix.governator.lifecycle.LifecycleManager;
 
 /**
@@ -26,7 +30,9 @@ public class ExampleEurekaGovernatedService {
 
     private static ExampleServiceBase init() throws Exception {
         System.out.println("Creating injector for Example Service");
-        LifecycleInjectorBuilder builder = LifecycleInjector.builder();
+
+        // Simulated Child Injectors required for the optional @Injects in DiscoveryClient
+        LifecycleInjectorBuilder builder = LifecycleInjector.builder().withMode(LifecycleInjectorMode.SIMULATED_CHILD_INJECTORS);
         builder.withModules(
                 new AbstractModule() {
                     @Override
@@ -36,6 +42,9 @@ public class ExampleEurekaGovernatedService {
                         // the default impl of EurekaInstanceConfig is CloudInstanceConfig, which we only want in an AWS
                         // environment. Here we override that by binding MyDataCenterInstanceConfig to EurekaInstanceConfig.
                         bind(EurekaInstanceConfig.class).to(MyDataCenterInstanceConfig.class);
+
+                        // (DiscoveryClient optional bindings) bind the optional event bus
+                        // bind(EventBus.class).to(EventBusImpl.class).in(Scopes.SINGLETON);
                     }
                 },
                 new EurekaModule(),
